refactor: api auth 관련 폴더 생성
